+++ sumo/src/dfrouter/RODFNet.cpp
@@ -294,7 +294,7 @@ RODFNet::computeRoutesFor(ROEdge* edge, RODFRouteDesc& base, int /*no*/,
         // ... else: loop over the next edges
         const ROEdgeVector& appr  = myApproachedEdges.find(last)->second;
         bool hadOne = false;
-        for (size_t i = 0; i < appr.size(); i++) {
+        for (int i = 0; i < appr.size(); i++) {
             if (find(current.edges2Pass.begin(), current.edges2Pass.end(), appr[i]) != current.edges2Pass.end()) {
                 // do not append an edge twice (do not build loops)
                 continue;
@@ -747,10 +747,10 @@ RODFNet::isSource(const RODFDetector& det, ROEdge* edge,
                 // the next is a hack for the A100 scenario...
                 //  We have to look into further edges herein edges
                 const ROEdgeVector& appr = myApproachingEdges.find(edge)->second;
-                size_t noOk = 0;
-                size_t noFalse = 0;
-                size_t noSkipped = 0;
-                for (size_t i = 0; i < appr.size(); i++) {
+                int noOk = 0;
+                int noFalse = 0;
+                int noSkipped = 0;
+                for (int i = 0; i < appr.size(); i++) {
                     if (!hasDetector(appr[i])) {
                         noOk++;
                     } else {
@@ -781,11 +781,11 @@ RODFNet::isSource(const RODFDetector& det, ROEdge* edge,
 
     // let's check the edges in front
     const ROEdgeVector& appr = myApproachingEdges.find(edge)->second;
-    size_t noOk = 0;
-    size_t noFalse = 0;
-    size_t noSkipped = 0;
+    int noOk = 0;
+    int noFalse = 0;
+    int noSkipped = 0;
     seen.push_back(edge);
-    for (size_t i = 0; i < appr.size(); i++) {
+    for (int i = 0; i < appr.size(); i++) {
         bool had = std::find(seen.begin(), seen.end(), appr[i]) != seen.end();
         if (!had) {
             if (isSource(det, appr[i], seen, detectors, strict)) {
@@ -867,9 +867,9 @@ RODFNet::isDestination(const RODFDetector& det, ROEdge* edge, ROEdgeVector& seen
     }
     const ROEdgeVector& appr  = myApproachedEdges.find(edge)->second;
     bool isall = true;
-    size_t no = 0;
+    int no = 0;
     seen.push_back(edge);
-    for (size_t i = 0; i < appr.size() && isall; i++) {
+    for (int i = 0; i < appr.size() && isall; i++) {
         bool had = std::find(seen.begin(), seen.end(), appr[i]) != seen.end();
         if (!had) {
             if (!isDestination(det, appr[i], seen, detectors)) {
@@ -917,7 +917,7 @@ RODFNet::isFalseSource(const RODFDetector& det, ROEdge* edge, ROEdgeVector& seen
 
     const ROEdgeVector& appr  = myApproachedEdges.find(edge)->second;
     bool isall = false;
-    for (size_t i = 0; i < appr.size() && !isall; i++) {
+    for (int i = 0; i < appr.size() && !isall; i++) {
         //printf("checking %s->\n", appr[i].c_str());
         bool had = std::find(seen.begin(), seen.end(), appr[i]) != seen.end();
         if (!had) {
