+++ sumo/src/guisim/GUIEdge.cpp
@@ -90,7 +90,7 @@ GUIEdge::~GUIEdge() {
 
 
 MSLane&
-GUIEdge::getLane(size_t laneNo) {
+GUIEdge::getLane(int laneNo) {
     assert(laneNo < myLanes->size());
     return *((*myLanes)[laneNo]);
 }
@@ -155,7 +155,7 @@ GUIEdge::getBoundary() const {
 
 void
 GUIEdge::fill(std::vector<GUIEdge*>& netsWrappers) {
-    size_t size = MSEdge::dictSize();
+    int size = MSEdge::dictSize();
     netsWrappers.reserve(size);
     for (DictType::iterator i = myDict.begin(); i != myDict.end(); ++i) {
         if (i->second->getPurpose() != MSEdge::EDGEFUNCTION_DISTRICT) {
@@ -195,7 +195,7 @@ GUIEdge::getParameterWindow(GUIMainWindow& app,
     ret->mkItem("mean vehicle speed [m/s]", true, new FunctionBinding<GUIEdge, SUMOReal>(this, &GUIEdge::getMeanSpeed));
     ret->mkItem("flow [veh/h/lane]", true, new FunctionBinding<GUIEdge, SUMOReal>(this, &GUIEdge::getFlow));
     ret->mkItem("routing speed [m/s]", true, new FunctionBinding<MSEdge, SUMOReal>(this, &MSEdge::getRoutingSpeed));
-    ret->mkItem("#vehicles", true, new CastingFunctionBinding<GUIEdge, SUMOReal, unsigned int>(this, &GUIEdge::getVehicleNo));
+    ret->mkItem("#vehicles", true, new CastingFunctionBinding<GUIEdge, SUMOReal, int>(this, &GUIEdge::getVehicleNo));
     ret->mkItem("vehicle ids", false, getVehicleIDs());
     // add segment items
     MESegment* segment = getSegmentAtPosition(parent.getPositionInformation());
@@ -206,7 +206,7 @@ GUIEdge::getParameterWindow(GUIMainWindow& app,
     ret->mkItem("segment brutto occupancy [%]", true, new FunctionBinding<MESegment, SUMOReal>(segment, &MESegment::getRelativeOccupancy, 100));
     ret->mkItem("segment mean vehicle speed [m/s]", true, new FunctionBinding<MESegment, SUMOReal>(segment, &MESegment::getMeanSpeed));
     ret->mkItem("segment flow [veh/h/lane]", true, new FunctionBinding<MESegment, SUMOReal>(segment, &MESegment::getFlow));
-    ret->mkItem("segment #vehicles", true, new CastingFunctionBinding<MESegment, SUMOReal, size_t>(segment, &MESegment::getCarNumber));
+    ret->mkItem("segment #vehicles", true, new CastingFunctionBinding<MESegment, SUMOReal, int>(segment, &MESegment::getCarNumber));
     ret->mkItem("segment leader leave time", true, new FunctionBinding<MESegment, SUMOReal>(segment, &MESegment::getEventTimeSeconds));
 
     // close building
@@ -305,7 +305,7 @@ GUIEdge::drawMesoVehicles(const GUIVisualizationSettings& s) const {
         // draw the meso vehicles
         vehicleControl->secureVehicles();
         AbstractMutex::ScopedLocker locker(myLock);
-        size_t laneIndex = 0;
+        int laneIndex = 0;
         MESegment::Queue queue;
         for (std::vector<MSLane*>::const_iterator msl = myLanes->begin(); msl != myLanes->end(); ++msl, ++laneIndex) {
             GUILane* l = static_cast<GUILane*>(*msl);
@@ -317,11 +317,11 @@ GUIEdge::drawMesoVehicles(const GUIVisualizationSettings& s) const {
                 if (laneIndex < segment->numQueues()) {
                     // make a copy so we don't have to worry about synchronization
                     queue = segment->getQueue(laneIndex);
-                    const size_t queueSize = queue.size();
+                    const int queueSize = queue.size();
                     SUMOReal vehiclePosition = segmentOffset + length;
                     // draw vehicles beginning with the leader at the end of the segment
                     SUMOReal xOff = 0;
-                    for (size_t i = 0; i < queueSize; ++i) {
+                    for (int i = 0; i < queueSize; ++i) {
                         GUIMEVehicle* veh = static_cast<GUIMEVehicle*>(queue[queueSize - i - 1]);
                         const SUMOReal vehLength = veh->getVehicleType().getLengthWithGap();
                         while (vehiclePosition < segmentOffset) {
@@ -347,13 +347,13 @@ GUIEdge::drawMesoVehicles(const GUIVisualizationSettings& s) const {
 
 
 
-unsigned int
+int
 GUIEdge::getVehicleNo() const {
-    size_t vehNo = 0;
+    int vehNo = 0;
     for (MESegment* segment = MSGlobals::gMesoNet->getSegmentForEdge(*this); segment != 0; segment = segment->getNextSegment()) {
         vehNo += segment->getCarNumber();
     }
-    return (unsigned int)vehNo;
+    return (int)vehNo;
 }
 
 
@@ -414,7 +414,7 @@ GUIEdge::setColor(const GUIVisualizationSettings& s) const {
 
 
 bool
-GUIEdge::setFunctionalColor(size_t activeScheme) const {
+GUIEdge::setFunctionalColor(int activeScheme) const {
     switch (activeScheme) {
         case 9: {
             const PositionVector& shape = getLanes()[0]->getShape();
@@ -430,7 +430,7 @@ GUIEdge::setFunctionalColor(size_t activeScheme) const {
 
 bool
 GUIEdge::setMultiColor(const GUIColorer& c) const {
-    const size_t activeScheme = c.getActive();
+    const int activeScheme = c.getActive();
     mySegmentColors.clear();
     switch (activeScheme) {
         case 10: // alternating segments
@@ -477,7 +477,7 @@ GUIEdge::setMultiColor(const GUIColorer& c) const {
 
 
 SUMOReal
-GUIEdge::getColorValue(size_t activeScheme) const {
+GUIEdge::getColorValue(int activeScheme) const {
     switch (activeScheme) {
         case 1:
             return gSelected.isSelected(getType(), getGlID());
@@ -501,7 +501,7 @@ GUIEdge::getColorValue(size_t activeScheme) const {
 
 
 SUMOReal
-GUIEdge::getScaleValue(size_t activeScheme) const {
+GUIEdge::getScaleValue(int activeScheme) const {
     switch (activeScheme) {
         case 1:
             return gSelected.isSelected(getType(), getGlID());
